Uses AI to switch signals in a 4way junction
This code is a simulation of a 4 way junction which uses AI to switch the signals instead of a normal timer. There are 2 AIs, one calculates bsed on the accurate distance of the cars from the signal. one calculates based on the distance rounded to the nearest 100px(a car is 24px x 16px).